// SMS.Web/src/api-client/api-client.ts
export interface RegisterDto {
  email?: string | undefined;
  firstName?: string | undefined;
  middleName?: string | undefined;
  lastName?: string | undefined;
  roles?: string[] | undefined;
}